+++ src/main/java/rocks/zipcode/atm/bank/PremiumAccount.java
@@ -12,7 +12,7 @@ public class PremiumAccount extends Account {
     }
 
     @Override
-    protected boolean canWithdraw(int amount) {
+    protected boolean canWithdraw(Float amount) {
 
         return getBalance() + OVERDRAFT_LIMIT >= amount;
     }
